+++ projects/ship/src/main/java/org/openmuc/jeebus/ship/node/ShipNodeImpl.java
@@ -27,15 +27,14 @@ import org.slf4j.LoggerFactory;
 
 import javax.net.ssl.SSLException;
 import java.net.InetSocketAddress;
-import java.net.URI;
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
-import java.util.Set;
+import java.net.URISyntaxException;
+import java.util.*;
 import java.util.concurrent.Executors;
 import java.util.concurrent.Future;
 import java.util.concurrent.ScheduledExecutorService;
 import java.util.concurrent.TimeUnit;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
 import static org.openmuc.jeebus.ship.node.KeyManagement.encodeSkiAsString;
 public class ShipNodeImpl {
@@ -131,9 +130,17 @@ public class ShipNodeImpl {
                 getOwnSki(),
                 server.getBoundSocketAddresses()
             );
-
         }
+    }
 
+    public Set<WebSocketHandler> getAllWebSocketHandlers() {
+        return Stream.concat(
+            servers.stream()
+                .map(ShipServer::getHandlers)
+                .flatMap(Collection::stream),
+            clients.stream()
+                .map(ShipClient::getHandler)
+        ).collect(Collectors.toSet());
     }
 
     public void closeDoubleConns(WebSocketHandler current) {
@@ -182,7 +189,7 @@ public class ShipNodeImpl {
         }
     }
 
-    public ShipClient createClient(URI serverUri) {
+    public ShipClient createClient(InetSocketAddress address, String path) {
         try {
             SslContext clientCtx = sslContextFactory.generateClientSslContext(
                 keyManagement.getCert()
@@ -194,7 +201,8 @@ public class ShipNodeImpl {
             nodeCtx.setConnHandler(connHandler);
             ShipClient client = new ShipClient(
                 clientCtx,
-                serverUri,
+                address,
+                path,
                 nodeCtx,
                 this
             );
@@ -205,7 +213,7 @@ public class ShipNodeImpl {
             log.error("Exception while creating a SHIP client: ", e);
             Thread.currentThread().interrupt();
         }
-        catch (SSLException e) {
+        catch (SSLException | URISyntaxException e) {
             log.error("Exception while creating a SHIP client: ", e);
         }
         return null;