+++ CHANGELOG.md
@@ -5,12 +5,20 @@
 ### Changed
 
 - Overhaul API
-   - introduce fluent builders
-   - introduce reasonable default configuration
+  - Introduce new configuration system:
+    - Builder pattern with fluent methods
+    - Reasonable default values
+    - Value validation
+    - Immutable config object
+  - Change `Ship#openConnection` to use `InetSocketAddress` and WSS path instead of
+    address-strings
 
 ### Added
 
+- Introduce ShipService class to represent SHIP mDNS Services
 - Fully support brand, type and serial mDNS TXT-Record flags
+- Implement a fallback to add jmDNS scope IDs to IPv6 link-local addresses that lack
+  network interface information
 
 ### Removed